Fix engine version semver comparison
Some VSCode extensions have an engine version requirement that includes
a pre-release identifier. E.g., as of writing, the "copilot-chat"
extension requires an engine version of "^1.106.0-20251103". The current
code incorrectly assumed that a VSCode version of "1.106.0" does _not_
match this requirement. However, it should, because the "-XXX" suffix
denotes a prerelease, and as such, a non-pre-release version satisfies
this requirement.

To fix this, we implement a proper semver-aware comparator that supports
comparing SemVers, including pre-release versions, accurately.